requirements.txt
langchain
langchain-openai
langchain-deepseek
pyyaml
httpx
pydantic
python-dotenv
fastapi
uvicorn
python-multipart
sqlalchemy
aiofiles
pytest
pytest-asyncio
